Added erlang push notification in submit method
With this commit, the state_changes controller will check whether the
PHP installation has the 'peb' module installed and the 'erlang.php'
file exists in the config directory. If yes, it'll send a
'hsapi_changed' atom to the registered process 'hsapi_listener' at the
node specified in the $erlang_node variable. This makes it possible to
develop push APIs interfacing HackSense.